fix: code cleanup and structure GetAllEmployees from CVPartner
