+++ src/Employee/Api/EmployeeController.cs
@@ -1,12 +1,6 @@
-using CvPartner.DTOs;
-
 using Employee.Repositories;
-
 using Microsoft.AspNetCore.Mvc;
-using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.Logging;
-
-using Refit;
 
 namespace Employee.Api;
 
@@ -16,34 +10,18 @@ public class EmployeeController : ControllerBase
 {
     private readonly EmployeeContext _db;
     private readonly ILogger<EmployeeController> _logger;
-    private readonly IConfiguration _congig;
-     public interface IEmployeeApi
-    {
-        [Get("/users?")]
-        Task<IEnumerable<CVPartnerUserDTO>> GetAllEmployee([Authorize("Token")] string authorization);
-    }
-
-    public EmployeeController(EmployeeContext db, ILogger<EmployeeController> logger, IConfiguration config)
+    
+    public EmployeeController(EmployeeContext db, ILogger<EmployeeController> logger)
     {
         _db = db;
         _logger = logger;
-        _congig = config;
     }
-
+    
     [HttpGet]
-    public async Task<IEnumerable<CVPartnerUserDTO>> Get()
+    public List<Models.Employee> Get()
     {
-        // _logger.LogInformation("Getting employees from database");
-        var apikey = _congig["Employee:ServiceApiKey"];
-
-        Console.WriteLine("HER");
-        Console.WriteLine("skal stå her: " + apikey);
-        var cvPartnerApi = RestService.For<IEmployeeApi>("https://variant.cvpartner.com/api/v1");
-        var employees = await cvPartnerApi.GetAllEmployee(apikey);
-
+        _logger.LogInformation("Getting employees from database");
+        var employees = _db.Employees.ToList();
         return employees;
-
-        // var employees = _db.Employees.ToList();
-        // return employees;
     }
 }
