+++ src/Shared/AppSettings.cs
@@ -4,4 +4,5 @@ public record AppSettings
 {
     public Uri AzureAppConfigUri { get; set; }
     public bool UseAzureAppConfig { get; set; }
+    public string Token { get; set; }
 }
