+++ src/WebApi/WebApi.csproj
@@ -23,4 +23,10 @@
     <ProjectReference Include="..\Employee\Employee.csproj" />
     <ProjectReference Include="..\Shared\Shared.csproj" />
   </ItemGroup>
+    
+  <ItemGroup>
+    <Content Update="appsettings.Local.json">
+      <CopyToPublishDirectory>Never</CopyToPublishDirectory>
+    </Content>
+  </ItemGroup>
 </Project>
