+++ src/WebApi/appsettings.json
@@ -1,7 +1,8 @@
 {
   "AppSettings": {
     "AzureAppConfigUri": "replace_me_on_deploy",
-    "UseAzureAppConfig": false
+    "UseAzureAppConfig": false,
+    "Token" : "replace_me_on_deploy"
   },
   "ConnectionStrings": {
     "EmployeeDatabase": "Server=127.0.0.1,1433;Database=EmployeeDatabase;User Id=sa;Password=yourStrong(!)Password"
